Moved git parent dir from MadgwickFusion... to present dir.
