+++ MadgwickFusionAdvancedJoystick/src/Fusion/CMakeLists.txt
