+++ MadgwickFusionAdvancedJoystick/src/Fusion/FusionCompass.h
