+++ MadgwickFusionAdvancedJoystick/src/Fusion/FusionOffset.h
