+++ test_MadgwickFusionAdvancedJoystick/test_utils.h
@@ -1,4 +1,4 @@
-#ifndef NO_TESTS
+#ifndef TEST_UTILS_H
 
 void test_set_calib_helper_vector();
 void test_set_calib_helper_vector_fail();
@@ -9,4 +9,4 @@ void test_set_calib_helper_vector_vector_fail();
 void test_set_calib_helper_vector_matrix();
 void test_set_calib_helper_vector_matrix_fail();
 
-#endif // NO_TESTS
+#endif // TEST_UTILS_H
